feat(website): build the docs site under typescript 7 (tsgo rc)
Moving the website's `typescript` to `catalog:typescript` bumps it to the
`7.0.1-rc` tsgo native preview, which the site's npm-`typescript` consumers do
not accept:

- `next` loads `next.config.ts` via `require("typescript")`, and the rc package
  ships no CommonJS entry (`No "exports" main defined`).
- `typedoc` does not support TypeScript 7.
- Nextra's shiki highlighter pulls in `twoslash`, which imports the classic
  `typescript` package and fails to resolve under the rc build.

Fixes, in order:

- Rename `next.config.ts` to `next.config.mjs` so Next never takes the
  TS-config load path, and set `typescript.ignoreBuildErrors` because Next's
  built-in type-check step also expects the classic npm compiler.
- Stub the `twoslash` module out of the webpack graph (the site uses no
  `twoslash` code blocks), preserving Nextra's own webpack setup.
- Drop typedoc entirely: the `typedoc` devDependency, the standalone
  `build:typedoc` script (never part of `build`), `build/typedoc.js`,
  `typedoc.json`, the `/api` sidebar entry in the docs `_meta.ts`, and the now
  unused `public/api/` and `typedoc-json/` gitignore lines.

The site carries no ttsc/typia transform in its own sources (its tsconfig
`plugins` holds only the Next plugin), so no bundler-level transform is needed.
