+++ website/.gitignore
@@ -3,14 +3,12 @@ _pagefind/
 node_modules/
 out/
 packages/
-public/api/
 public/blog/rss.xml
 public/compiler/
 public/robots.txt
 public/sitemap*.xml
 src/raw/
 tsconfig.tsbuildinfo
-typedoc-json/
 
 src/compiler/PlaygroundExampleStorage.ts
 !.env
