--- website/next.config.ts
@@ -1,18 +1,0 @@
-import nextra from "nextra";
-
-const withNextra = nextra({
-  // ... Other Nextra config options
-});
-
-// You can include other Next.js configuration options here, in addition to Nextra settings:
-export default withNextra({
-  // ... Other Next.js config options
-  output: "export",
-  trailingSlash: true,
-  eslint: {
-    ignoreDuringBuilds: true,
-  },
-  images: {
-    unoptimized: true,
-  },
-});
