--- website/typedoc.json
@@ -1,12 +1,0 @@
-{
-  "entryPointStrategy": "packages",
-  "entryPoints": [
-    "../packages/typia",
-    "../packages/interface",
-    "../packages/utils",
-    "../packages/mcp",
-    "../packages/langchain",
-    "../packages/vercel",
-  ],
-  "out": "./public/api"
-}
